Add IRS 2004 life table
This has more data that the previous life table I was using.
It allows me to calculate joint probability.

Added lognormal monte carlo.

All in an attempt to replicate the Blanchett results…which I still
can’t do :(

For a 5.5% constant withdrawal and a 40% portfolio, I’m seeing a WER of
around 53% but it should be well into the 60s.

I’m not sure why I’m seeing such a large discrepancy….