CEW 0-value fixes
CEW (and by extension WER) behaves poorly if there are any 0s
in the cashflow list. A single zero means CEW always returns 0.
That skews results a lot. To counter that, Blanchett et al add a tiny
amount of cashflow every year.

This adds an automatic fudge factor to WER and an assertion to
CEW to check for 0s.