My mortality calculations were totally wrong
Which is what accounted for a lot of the discrepancy between
my WER results and Blanchett et al’s WER results. (I was making them
die younger, which causes a bigger gap between the perfect information
Sustainable Spending Rate.)

Also, found some issues with the lifetable but that’s a future commit.