Totally redo how mortality works
- Most important: fixed the calculations so people aren’t dying so
early.
- Added 2 other life tables (both of which have people living longer
than the original one I was using)

With this change, my WER calculations FINALLY agree with those in
Blanchett et al.

I thought my monte carlo simulation of market returns was the culprit
(and it was, a bit) but the real problem was the mortality calculations
were totally wrong :(