A totally different way to generate monte carlo returns
This comes from some book on finance programming for Python. It seems
massively more complicated than what I’m currently doing. I’m not sure
it is worth it.

However, it does some interesting things with vectorization of numpy
that were new to me, so I’ll keep it around for inspiration.