Refactor withdrawal harness
The mixture of "things at the beginning of the year" (like withdrawal) and "things at the end of the year" (like returns) was unclear. Did the portfolio value refer to the before the withdrawal? After the withdrawal but before the yearly returns?

We also weren't tracking "real returns".

Probably lots of existing notebooks are broken now, because the AnnualResults tuple is used EVERYWHERE.