revision to simulate monthly bonds
Instead of having totally separate annual and monthly code paths, let's try just having one.

Instead of thinking of it as annual or monthly just think of it as one abstract "time tick".  That means that yields need to be converted to monthly yields. And the yield curve needs steps for each month along the curve.